properly clean up kernel pointers, small refactor
